requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.0.1
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.4
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
pytest==8.3.3
python-dotenv==1.0.1
PyYAML==6.0.2
setuptools==75.1.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
wheel==0.44.0
python-jose==3.3.0
PyJWT==2.10.0
pytz==2024.2
distro==1.9.0
jiter==0.8.0
openai==1.55.3
tqdm==4.67.1
charset-normalizer==3.4.0
discord-webhook==1.3.1
requests==2.32.3
urllib3==2.2.3